server/pyproject.toml
[project]
name = "python_chatbot"
version = "0.1.0"
description = "Fullstack application with Python to create a chatbot using LLMs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.15.0",
    "pydantic-settings>=2.9.1",
]

[tool.ruff.format]
quote-style = 'single'

[dependency-groups]
dev = ["ipython>=8.18.1"]
